refactor: Section 추가, 삭제 책임을 LineService가 갖도록 변경
