+++ src/main/java/subway/controller/LineController.java
@@ -1,11 +1,9 @@
 package subway.controller;
 
 import java.net.URI;
-import java.sql.SQLException;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,6 +13,8 @@ import org.springframework.web.bind.annotation.RestController;
 import subway.application.LineService;
 import subway.controller.dto.LineRequest;
 import subway.controller.dto.LineResponse;
+import subway.controller.dto.SectionCreateRequest;
+import subway.controller.dto.SectionDeleteRequest;
 
 @RestController
 @RequestMapping("/lines")
@@ -48,8 +48,16 @@ public class LineController {
         return ResponseEntity.noContent().build();
     }
 
-    @ExceptionHandler(SQLException.class)
-    public ResponseEntity<Void> handleSQLException() {
-        return ResponseEntity.badRequest().build();
+    @PostMapping("/{id}/sections")
+    public ResponseEntity<Void> createSection1(@PathVariable Long id, @RequestBody SectionCreateRequest request) {
+        lineService.createSection(id, request);
+        return ResponseEntity.created(URI.create("/sections")).build();
+    }
+
+    @DeleteMapping("/{id}/sections")
+    public ResponseEntity<Void> deleteSection2(@PathVariable Long id,
+                                               @RequestBody SectionDeleteRequest request) {
+        lineService.deleteSection(id, request);
+        return ResponseEntity.noContent().build();
     }
 }
